:bug: Solve the problem of bounty return type error in PlayerFinderAdapter.

+++ utopia-gamification/src/main/kotlin/tw/waterballsa/utopia/utopiagamification/repositories/PlayerFinderAdapter.kt
@@ -15,6 +15,6 @@ class PlayerFinderAdapter(
         return playerRepository.findPlayerById(id)?.toMiniGamePlayer()
     }
 
-    private fun Player.toMiniGamePlayer(): MiniGamePlayer = MiniGamePlayer(id,  bounty.toUInt())
+    private fun Player.toMiniGamePlayer(): MiniGamePlayer = MiniGamePlayer(id,  bounty)
 
 }
